pyproject.toml
[project]
name = "robo-to-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.1",
    "python-dotenv>=1.1.0",
    "robocorp>=3.0.0",
    "rpaframework>=30.0.2",
]
